+++ lambdas/src/controllers/routes.ts
@@ -10,6 +10,7 @@ import { getEmotesByOwnerHandler, getEmotesHandler } from '../apis/collections/c
 import { getWearablesByOwnerHandler, getWearablesHandler } from '../apis/collections/controllers/wearables'
 import { OffChainWearablesManager } from '../apis/collections/off-chain/OffChainWearablesManager'
 import { getContents, getInfo, getScenes } from '../apis/content-v2/controllers/translator'
+import { getCatalystServersList, getDenylistedNamesList, getPOIsList } from '../apis/contracts/controllers/contracts'
 import { Bean, Environment, EnvironmentConfig } from '../Environment'
 import { createFetchComponent } from '../ports/fetcher'
 import { createThirdPartyAssetFetcher, ThirdPartyAssetFetcher } from '../ports/third-party/third-party-fetcher'
@@ -17,7 +18,6 @@ import { DAOCache } from '../service/dao/DAOCache'
 import { SmartContentClient } from '../utils/SmartContentClient'
 import { SmartContentServerFetcher } from '../utils/SmartContentServerFetcher'
 import { TheGraphClient } from '../utils/TheGraphClient'
-import { getCatalystServersList, getDenylistedNamesList, getPOIsList } from './handlers/contracts/handlers'
 import { validateSignature } from './handlers/crypto/handlers'
 import { hotScenes, realmsStatus } from './handlers/explorer/handlers'
 import { getResizedImage } from './handlers/images/handlers'
